+++ operate/client/src/App/Dashboard/PartiallyExpandableDataTable/index.tsx
@@ -9,11 +9,11 @@
 import React from 'react';
 import {
   DataTable,
-  TableBody,
   TableContainer,
   TableExpandHeader,
   TableHeader,
   TableRow,
+  type DataTableRenderProps,
 } from '@carbon/react';
 import {
   Table,
@@ -22,22 +22,28 @@ import {
   TableExpandedRow,
   TableHead,
 } from './styled';
+import {InfiniteScroller} from 'modules/components/InfiniteScroller';
 
-/**
- * Non-paginated variant of PartiallyExpandableDataTable.
- * Used for simple tables with all data loaded at once (e.g., IncidentsByError).
- *
- * For paginated tables with infinite scroll, use v2/PartiallyExpandableDataTable instead.
- * Will be removed in the scope of #44728
- */
+type Row = {
+  id: string;
+  [key: string]: React.ReactNode;
+};
+
 type Props = {
   headers: {key: string; header: string; width?: string}[];
-  rows: React.ComponentProps<typeof DataTable>['rows'];
+  rows: Row[];
   className?: string;
   expandedContents?: {
     [key: string]: React.ReactElement<{tabIndex: number}>;
   };
   dataTestId?: string;
+  onVerticalScrollStartReach?: React.ComponentProps<
+    typeof InfiniteScroller
+  >['onVerticalScrollStartReach'];
+  onVerticalScrollEndReach?: React.ComponentProps<
+    typeof InfiniteScroller
+  >['onVerticalScrollEndReach'];
+  scrollableContainerRef: React.RefObject<HTMLElement | null>;
 };
 
 const PartiallyExpandableDataTable: React.FC<Props> = ({
@@ -45,10 +51,63 @@ const PartiallyExpandableDataTable: React.FC<Props> = ({
   rows,
   expandedContents,
   dataTestId,
+  onVerticalScrollStartReach,
+  onVerticalScrollEndReach,
+  scrollableContainerRef,
 }) => {
+  const tableBodyContent = (
+    rows: DataTableRenderProps<Row, React.ReactNode[]>['rows'],
+    headers: DataTableRenderProps<Row, React.ReactNode[]>['headers'],
+    getRowProps: DataTableRenderProps<Row, React.ReactNode[]>['getRowProps'],
+    getExpandedRowProps: DataTableRenderProps<
+      Row,
+      React.ReactNode[]
+    >['getExpandedRowProps'],
+  ) => {
+    return rows.map((row, index: number) => {
+      const expandedContent = expandedContents?.[row.id];
+
+      const isExpandable =
+        expandedContent !== undefined && React.isValidElement(expandedContent);
+
+      const {key, ...props} = getRowProps({row});
+      return (
+        <React.Fragment key={row.id}>
+          <TableExpandRow
+            {...props}
+            data-testid={`${dataTestId}-${index}`}
+            $isExpandable={isExpandable}
+          >
+            {row.cells.map((cell) => (
+              <ExpandableTableCell key={cell.id}>
+                {cell.value}
+              </ExpandableTableCell>
+            ))}
+          </TableExpandRow>
+
+          {isExpandable && (
+            <TableExpandedRow
+              colSpan={headers.length + 1}
+              {...getExpandedRowProps({row})}
+            >
+              {row.isExpanded
+                ? React.cloneElement(expandedContent, {
+                    tabIndex: 0,
+                  })
+                : null}
+            </TableExpandedRow>
+          )}
+        </React.Fragment>
+      );
+    });
+  };
+
   return (
-    <DataTable size="sm" headers={headers} rows={rows}>
-      {({
+    <DataTable<Row, React.ReactNode[]>
+      size="sm"
+      headers={headers}
+      rows={rows}
+      children={({
         rows,
         headers,
         getTableContainerProps,
@@ -74,49 +133,24 @@ const PartiallyExpandableDataTable: React.FC<Props> = ({
                 })}
               </TableRow>
             </TableHead>
-            <TableBody>
-              {rows.map((row, index) => {
-                const expandedContent = expandedContents?.[row.id];
-
-                const isExpandable =
-                  expandedContent !== undefined &&
-                  React.isValidElement(expandedContent);
-
-                const {key, ...props} = getRowProps({row});
-                return (
-                  <React.Fragment key={row.id}>
-                    <TableExpandRow
-                      {...props}
-                      data-testid={`${dataTestId}-${index}`}
-                      $isExpandable={isExpandable}
-                    >
-                      {row.cells.map((cell) => (
-                        <ExpandableTableCell key={cell.id}>
-                          {cell.value}
-                        </ExpandableTableCell>
-                      ))}
-                    </TableExpandRow>
-
-                    {isExpandable && (
-                      <TableExpandedRow
-                        colSpan={headers.length + 1}
-                        {...getExpandedRowProps({row})}
-                      >
-                        {row.isExpanded
-                          ? React.cloneElement(expandedContent, {
-                              tabIndex: 0,
-                            })
-                          : null}
-                      </TableExpandedRow>
-                    )}
-                  </React.Fragment>
-                );
-              })}
-            </TableBody>
+            <InfiniteScroller
+              onVerticalScrollStartReach={onVerticalScrollStartReach}
+              onVerticalScrollEndReach={onVerticalScrollEndReach}
+              scrollableContainerRef={scrollableContainerRef}
+            >
+              <tbody>
+                {tableBodyContent(
+                  rows,
+                  headers,
+                  getRowProps,
+                  getExpandedRowProps,
+                )}
+              </tbody>
+            </InfiniteScroller>
           </Table>
         </TableContainer>
       )}
-    </DataTable>
+    />
   );
 };
